Add files via upload
In numerical methods, there exists different methods of solving a differential equation with an initial condition. In nature there are types of differential equations that are difficult to find an exact solution, in our case we look at functional differential equations where the order of derivative is an integer between 0 and 1. We developed a numerical method called Adams to find numerical solutions for such equations.